+++ src/compiler/headers/variable/VariableInteger.h
@@ -8,10 +8,10 @@
 
 class VariableInteger final: public Variable {
 private:
-    int value;
+    long long value;
 
 public:
-    explicit VariableInteger(int value);
+    explicit VariableInteger(long long value);
 
     Variable* operator + (const Variable&) const override;
     Variable* operator - (const Variable&) const override;
@@ -19,7 +19,7 @@ public:
     Variable* operator / (const Variable&) const override;
 
     [[nodiscard]] bool getBool() const override;
-    [[nodiscard]] int getInteger() const override;
-    [[nodiscard]] double getDouble() const override;
+    [[nodiscard]] long long getInteger() const override;
+    [[nodiscard]] long double getDouble() const override;
     [[nodiscard]] std::string getString() const override;
 };
